fix!: correctly dump all schematics
Previously the `oneof` schematic could not correctly dump `map` values,
as it wasn't capable of knowing which schematic should be used to dump.

Now, the "dump" concept is just an inversion of the `unify` concept. The
unify, raw, and transform functions now receive a second parameter of
the direction of the unification, `:to` or `:from` (probably can come up
with better names).

This also means that values will be validated going to and from the user
data structure. So if you have a map from your domain model that you
want to dump to a plain outside data, it will validate the data as it
goes out so you aren't sending out bad data.