fix!: correctly handle required map keys with nil schematics (#39)
This patch remedies a bug where a required map key could have a nil
schematic, meaning set as the literal nil, but would be treated as if
the key was optional, rather than specifying the value is nil exactly.

This had downstream effects when used inside another schematic like the
`oneof` schematic. If you had two map schematics as options, both with
`nil` as the value schematic for its key, it would assume the first map
unified, but was just missing its key.

Now, map schematics with missing required keys will return errors with
"is missing".